Extracted a shared helper from the CSV stream response
Pure refactor with no behaviour change: the header wiring, the
no-transform cache directive and the pipeline teardown move from
stream-csv-response into a generic stream-response helper, with the
CSV variant becoming a thin wrapper that pins its content type and
its exact missing-filename error message. An upcoming zip download
needs the identical response plumbing, and duplicating it would let
the two copies drift.
